Fix unpacking multi-spec toolchains built with `--targets=all`.
When packing a toolchain archive, the specs and targets are specified
independently, so typically each spec must use the same list of targets;
but the special target `all` is expanded to a per-spec list of targets
during packing.  During unpacking, Romt had been detecting the union of
all targets present in the archive and applying this set of targets to
all detected specs in the archive, causing problems if one spec
supported more targets than the others.  Now Romt detects archives that
contain all targets for all included specs and converts back to the
special `all` target for verification and further processing.

See https://github.com/drmikehenry/romt/issues/17.